requirements.txt
absl-py==2.3.1
altair==5.5.0
astunparse==1.6.3
attrs==26.1.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
chardet==3.0.4
charset-normalizer==3.4.6
click==8.1.8
colorama==0.4.6
contourpy==1.3.0
cssselect==1.3.0
cycler==0.12.1
facenet-pytorch==2.5.3
feedfinder2==0.0.4
feedparser==6.0.12
filelock==3.19.1
flatbuffers==25.12.19
fonttools==4.60.2
fsspec==2025.10.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-pasta==0.2.0
googletrans==4.0.0rc1
grpcio==1.78.0
h11==0.9.0
h2==3.2.0
h5py==3.14.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface_hub==0.36.2
hyperframe==5.2.0
idna==2.10
importlib_metadata==8.7.1
importlib_resources==6.5.2
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.10.0
kiwisolver==1.4.7
langdetect==1.0.9
libclang==18.1.1
lxml==6.0.2
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.9.4
mdurl==0.1.2
mediapipe==0.10.33
ml_dtypes==0.5.4
mpmath==1.3.0
namex==0.1.0
narwhals==2.18.0
networkx==3.2.1
newspaper3k==0.2.8
nltk==3.9.2
numpy==1.26.4
opencv-contrib-python==4.8.0.74
opencv-python==4.8.0.74
opt_einsum==3.4.0
optree==0.19.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
plotly==6.6.0
protobuf==6.33.6
psutil==7.2.2
pyarrow==21.0.0
pycparser==2.23
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.36.2
regex==2026.1.15
requests==2.32.5
requests-file==3.0.1
rfc3986==1.5.0
rich==14.3.3
rpds-py==0.27.1
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
seaborn==0.13.2
sgmllib3k==1.0.0
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
sounddevice==0.5.5
soupsieve==2.8.3
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
textblob==0.19.0
tf_keras==2.20.1
threadpoolctl==3.6.0
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.22.2
toml==0.10.2
torch==2.8.0
torchvision==0.23.0
tornado==6.5.5
tqdm==4.67.3
transformers==4.57.6
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
vaderSentiment==3.3.2
watchdog==6.0.0
Werkzeug==3.1.6
wrapt==2.1.2
zipp==3.23.0
